motors control done but not polished
motors can be given an angle and they go to it. they can also have a
varied speed. there is still a LOT of functionality left to implement
int he motor class but you can move the motors to angles. so meh. not
that the 'centre all motors' method is considered the zero angle point
for kinematic calculations. this class would benefit from a "go here in
this time" method. but i imagine that is taken care of by the
kinematics.